rearranged EntryContentBlock component
