+++ components/EntryContentBlock/EntryContentBlock.style.js
@@ -18,3 +18,12 @@ export const ContentHeadline = styled.h2`
   left: 50%;
   transform: translate(-50%, -50%);
 `;
+
+export const EntryInput = styled.textarea`
+  width: 100%;
+  font-size: 1em;
+  margin-top: 15px;
+  font-family: sans-serif;
+  ${(prop) => (prop.long ? "height: 7em" : "")}
+  ${(prop) => (prop.short ? "height: 1.5em" : "")}
+`;
